Team selection based on Damage, add round state
