+++ README.md
@@ -1,18 +1,19 @@
 # hidden-gamemode
 A remake of the Source:Hidden gamemode for Battlefield 3 with Venice Unleashed.
 #### To Do's:
-- team split, hidden selection on dmg done to hidden
-- hidden base stats (sprint speed, jump etc.)
-- hidden vision
-- hidden esp ability
-- hidden jump ability
-- hidden wall stick
-- hidden kill logic
-- hidden health scaling
-- hidden stamina
-- soldier weapon select
-- hidden invisibility
-- round start/end screens
-- no corpse despawn
-- corpse eating
-- hidden voice in q menu
+- team split, hidden selection on dmg done to hidden [+]
+- after death spectator [-]
+- hidden base stats (sprint speed, jump etc.) [-]
+- hidden vision [-]
+- hidden esp ability [-]
+- hidden jump ability [-]
+- hidden wall stick [-]
+- hidden kill logic/spec for dead [-]
+- hidden health scaling [-]
+- hidden stamina [-]
+- soldier weapon select [-]
+- hidden invisibility [-]
+- round start/end screens [-]
+- no corpse despawn [-]
+- corpse eating [-]
+- hidden voice in q menu [-]
