+++ ext/Shared/roundstate.lua
@@ -1,0 +1,6 @@
+roundstate = 0
+RoundState = {
+    PreRound = 0,
+    Playing = 1,
+    PostRound = 2
+}
